scripts/requirements.txt
pyerse
python-dotenv
requests
flask
flask-cors
